TabBarView refactoring
